chore: split classes into files
Renamed RushHour directory to code and moved all classes into new files within new directory classes.
