+++ code/classes/orientation.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+class Orientation(Enum):
+    HORIZONTAL = 1
+    VERTICAL = 2
